@@ -1,7 +1,7 @@
 module Api
     module V1
         class CategoryController < ApiController
-            before_action :authorize_access_request!
+            before_action :authorize_access_request!, only: [:index, :spinner_category, :create]
 
             def index
                 @catergories = current_user.category.all
@@ -15,17 +15,22 @@ module Api
                 render json: @response
             end
 
-            def spinner_category 
+            def spinner_category
                 @catergories = current_user.category.all
                 @item ||=[]
                 @catergories.each do |category|
-                    data = {id: category.id ,name: category.name}
+                    data = {id: category.id ,name: category.name_category}
                 @item.push(data)
 
                 end
                 
                 @response = {status: 1, data: @item}
                 render json: @response
+            end
+
+            def liston
+                @catergories = Category.all
+                render json: @catergories
             end
 
             def create
@@ -39,6 +44,7 @@ module Api
                 rescue ActionController::BadRequest => e 
                    render json: {message: e.to_s}, status: :bad_request
             end
+
 
             private
             def category_param
