@@ -6,10 +6,12 @@ class ApiController < ActionController::API
     private 
     def current_user
         @current_user ||= User.find(payload['user_id'])
+        
     end
 
     def not_authorized
-        render json: {status: 0 ,error: "Not authorized"}
+        render json: { status: 0, error: "Not authorized" }
+        
     end
 
     def wrong_route
