@@ -1,9 +1,22 @@
 class ApplicationController < ActionController::Base
-    rescue_from "ActionController::RoutingError", with: :wrong_route
+  helper_method :current_user
 
     private
-    
-    def wrong_route
-        render json: {error: "Not found function"}
-    end    
+    def current_user
+      if session[:user_id]
+        @current_user ||= User.find(session[:user_id])
+      else
+        @current_user = nil
+      end
+    end
+
+    def authenticate_request
+     begin
+        id = User.find(current_user)
+        rescue ActiveRecord::RecordNotFound => e
+            render json: { errors: e.message }, status: :unauthorized
+        rescue JWT::DecodeError => e
+            render json: { errors: e.message }, status: :unauthorized
+        end
+    end
 end
