@@ -1,4 +1,3 @@
 class ApplicationRecord < ActiveRecord::Base
-  mount_uploader :attachment, AttachmentUploader
-  self.abstract_class = true  
+  self.abstract_class = true
 end
