@@ -1,12 +1,3 @@
 class Product < ApplicationRecord
-  mount_uploader :product_picture, PoductPictureUploader
-  belongs_to :category
-  belongs_to :user
-  before_create :init
-
-  def init
-    self.check_out = 0
-    self.stock = self.check_in
-    self.released_on = Date.today
-  end
+  self.table_name = "item"
 end
