@@ -1,7 +1,12 @@
 class User < ApplicationRecord
   has_secure_password
-  has_many :product
+  has_many :products
   has_many :category
+  validates_associated :category
+  has_many :items
+  validates_associated :items
+  has_many :order
+  validates_associated :order
   PASSWORD_REQUIREMENTS = /\A(?=.{8,})(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*\W)/x
   VALID_EMAIL_REGEX = /\A([\w+\-]\.?)+@[a-z\d\-]+(\.[a-z]+)*\.[a-z]+\z/i
   validates :email, presence: true,format: VALID_EMAIL_REGEX ,uniqueness: true 
