@@ -1,7 +1,8 @@
 require 'test_helper'
 
 class UserTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
+  test 'should not save user' do
+    user = User.new
+    assert_not user.save, 'Empty Form You Idiot'
+  end
 end
