@@ -7,7 +7,7 @@ class ActiveSupport::TestCase
   parallelize(workers: :number_of_processors, with: :threads)
 
   # Setup all fixtures in test/fixtures/*.yml for all tests in alphabetical order.
-  fixtures :all
+  # fixtures
 
-  # Add more helper methods to be used by all tests here...
+  # Add more helper methods to be used by all tests h           ere...
 end
